+++ libs/react/shelter-operator/src/lib/components/Schedule/ScheduleForm.tsx
@@ -5,105 +5,13 @@ import {
 } from '@monorepo/react/shelter';
 import { useState } from 'react';
 import { Button } from '../base-ui/buttons';
-import {
-  ScheduleExceptionsForm,
-  buildDefaultExceptionEntry,
-} from './ScheduleExceptionsForm';
+import { ScheduleExceptionsForm } from './ScheduleExceptionsForm';
 import type { ExceptionEntry, WeeklyFormState } from './types';
+import { buildScheduleInputs, hydrateExceptions, hydrateWeekly } from './utils';
 import {
   WeeklyScheduleEditor,
   buildDefaultWeeklyState,
 } from './WeeklyScheduleEditor';
-import { ORDERED_DAYS } from './WeeklyScheduleEditor/DayRow';
-
-export type { WeeklyFormState };
-
-// ─── State → GraphQL input ────────────────────────────────────────────────────
-
-function buildScheduleInputs(
-  scheduleType: ScheduleTypeChoices,
-  weekly: WeeklyFormState,
-  exceptions: ExceptionEntry[]
-): ScheduleInput[] {
-  const inputs: ScheduleInput[] = [];
-
-  for (const { value: day } of ORDERED_DAYS) {
-    const { open, startTime, endTime } = weekly[day];
-    if (!open || !startTime || !endTime) continue;
-
-    inputs.push({
-      scheduleType,
-      days: [day],
-      startTime: startTime.length === 5 ? `${startTime}:00` : startTime,
-      endTime: endTime.length === 5 ? `${endTime}:00` : endTime,
-      isException: false,
-    });
-  }
-
-  for (const ex of exceptions) {
-    if (!ex.startDate) continue;
-
-    inputs.push({
-      scheduleType,
-      startDate: ex.startDate,
-      endDate: ex.endDate || ex.startDate,
-      startTime: ex.closedAllDay
-        ? undefined
-        : ex.startTime.length === 5
-        ? `${ex.startTime}:00`
-        : ex.startTime || undefined,
-      endTime: ex.closedAllDay
-        ? undefined
-        : ex.endTime.length === 5
-        ? `${ex.endTime}:00`
-        : ex.endTime || undefined,
-      condition: ex.condition,
-      isException: true,
-    });
-  }
-
-  return inputs;
-}
-
-// ─── GraphQL → form state ────────────────────────────────────────────────────
-
-/** Strip seconds from "HH:MM:SS" → "HH:MM", pass through "HH:MM", return "" for nullish. */
-function toHHMM(time: string | null | undefined): string {
-  if (!time) {
-    return '';
-  }
-
-  return String(time).slice(0, 5);
-}
-
-function hydrateExceptions(schedules: ScheduleType[]): ExceptionEntry[] {
-  let _id = 0;
-
-  return schedules.map((s) => ({
-    localId: s.id ?? String(++_id),
-    startDate: s.startDate ?? '',
-    endDate: s.endDate ?? s.startDate ?? '',
-    closedAllDay: !s.startTime && !s.endTime,
-    startTime: toHHMM(s.startTime),
-    endTime: toHHMM(s.endTime),
-    condition: s.condition ?? undefined,
-  }));
-}
-
-function hydrateWeekly(schedules: ScheduleType[]): WeeklyFormState {
-  const state = buildDefaultWeeklyState();
-  for (const entry of schedules) {
-    if (!entry.day) continue;
-    state[entry.day] = {
-      open: true,
-      startTime: toHHMM(entry.startTime),
-      endTime: toHHMM(entry.endTime),
-    };
-  }
-  return state;
-}
-
-// ─── Component ────────────────────────────────────────────────────────────────
 
 type TProps = {
   scheduleType: ScheduleTypeChoices;
@@ -138,6 +46,7 @@ export function ScheduleForm(props: TProps) {
         <h3 className="text-base font-semibold text-gray-800 mb-4">
           Weekly Hours
         </h3>
+
         <WeeklyScheduleEditor value={weekly} onChange={setWeekly} />
       </section>
 
@@ -151,6 +60,7 @@ export function ScheduleForm(props: TProps) {
         <Button variant="floating" onClick={handleSave}>
           Save Schedule
         </Button>
+
         {onCancel && (
           <button
             type="button"
@@ -164,6 +74,3 @@ export function ScheduleForm(props: TProps) {
     </div>
   );
 }
-
-// Re-export for convenience
-export { buildDefaultExceptionEntry };
