+++ libs/react/shelter-operator/src/lib/components/Schedule/WeeklyScheduleEditor/DayRow.tsx
@@ -60,15 +60,3 @@ export function DayRow(props: TProps) {
     </div>
   );
 }
-
-// ─── Ordered display list ─────────────────────────────────────────────────────
-
-export const ORDERED_DAYS: { value: DayOfWeekChoices; label: string }[] = [
-  { value: DayOfWeekChoices.Monday, label: 'Monday' },
-  { value: DayOfWeekChoices.Tuesday, label: 'Tuesday' },
-  { value: DayOfWeekChoices.Wednesday, label: 'Wednesday' },
-  { value: DayOfWeekChoices.Thursday, label: 'Thursday' },
-  { value: DayOfWeekChoices.Friday, label: 'Friday' },
-  { value: DayOfWeekChoices.Saturday, label: 'Saturday' },
-  { value: DayOfWeekChoices.Sunday, label: 'Sunday' },
-];
