+++ libs/react/shelter-operator/src/lib/components/Schedule/WeeklyScheduleEditor/index.tsx
@@ -1,4 +1,4 @@
-export { DayRow, ORDERED_DAYS } from './DayRow';
+export { DayRow } from './DayRow';
 export { TimeRangeInput } from './TimeRangeInput';
 export { buildDefaultWeeklyState } from './utils';
 export { WeeklyScheduleEditor } from './WeeklyScheduleEditor';
