+++ libs/react/shelter-operator/src/lib/components/Schedule/constants.ts
@@ -1,0 +1,11 @@
+import { DayOfWeekChoices } from '@monorepo/react/shelter';
+
+export const ORDERED_DAYS: { value: DayOfWeekChoices; label: string }[] = [
+  { value: DayOfWeekChoices.Monday, label: 'Monday' },
+  { value: DayOfWeekChoices.Tuesday, label: 'Tuesday' },
+  { value: DayOfWeekChoices.Wednesday, label: 'Wednesday' },
+  { value: DayOfWeekChoices.Thursday, label: 'Thursday' },
+  { value: DayOfWeekChoices.Friday, label: 'Friday' },
+  { value: DayOfWeekChoices.Saturday, label: 'Saturday' },
+  { value: DayOfWeekChoices.Sunday, label: 'Sunday' },
+];
